Isolate tagged tests
Move functions used by more than one tag to a common file
