+++ govcd/common_test.go
@@ -7,11 +7,14 @@
 package govcd
 
 import (
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"net"
 	"net/http"
+	"net/url"
 	"regexp"
+	"sort"
 	"strconv"
 	"time"
 
@@ -253,4 +256,408 @@ func isTcpPortOpen(host, port string, timeout int) bool {
 		}
 	}
 
+}
+
+// moved from vapp_test.go
+func createVappForTest(vcd *TestVCD, vappName string) (*VApp, error) {
+	// Populate OrgVDCNetwork
+	var networks []*types.OrgVDCNetwork
+	net, err := vcd.vdc.GetOrgVdcNetworkByName(vcd.config.VCD.Network.Net1, false)
+	if err != nil {
+		return nil, fmt.Errorf("error finding network : %s", err)
+	}
+	networks = append(networks, net.OrgVDCNetwork)
+	// Populate Catalog
+	cat, err := vcd.org.GetCatalogByName(vcd.config.VCD.Catalog.Name, false)
+	if err != nil || cat == nil {
+		return nil, fmt.Errorf("error finding catalog : %s", err)
+	}
+	// Populate Catalog Item
+	catitem, err := cat.GetCatalogItemByName(vcd.config.VCD.Catalog.CatalogItem, false)
+	if err != nil {
+		return nil, fmt.Errorf("error finding catalog item : %s", err)
+	}
+	// Get VAppTemplate
+	vAppTemplate, err := catitem.GetVAppTemplate()
+	if err != nil {
+		return nil, fmt.Errorf("error finding vapptemplate : %s", err)
+	}
+	// Get StorageProfileReference
+	storageProfileRef, err := vcd.vdc.FindStorageProfileReference(vcd.config.VCD.StorageProfile.SP1)
+	if err != nil {
+		return nil, fmt.Errorf("error finding storage profile: %s", err)
+	}
+	// Compose VApp
+	task, err := vcd.vdc.ComposeVApp(networks, vAppTemplate, storageProfileRef, vappName, "description", true)
+	if err != nil {
+		return nil, fmt.Errorf("error composing vapp: %s", err)
+	}
+	// After a successful creation, the entity is added to the cleanup list.
+	// If something fails after this point, the entity will be removed
+	AddToCleanupList(vappName, "vapp", "", "createTestVapp")
+	err = task.WaitTaskCompletion()
+	if err != nil {
+		return nil, fmt.Errorf("error composing vapp: %s", err)
+	}
+	// Get VApp
+	vapp, err := vcd.vdc.GetVAppByName(vappName, true)
+	if err != nil {
+		return nil, fmt.Errorf("error getting vapp: %s", err)
+	}
+
+	err = vapp.BlockWhileStatus("UNRESOLVED", vapp.client.MaxRetryTimeout)
+	if err != nil {
+		return nil, fmt.Errorf("error waiting for created test vApp to have working state: %s", err)
+	}
+
+	return vapp, nil
+}
+
+// Checks whether an independent disk is attached to a VM, and detaches it
+// moved from disk_test.go
+func (vcd *TestVCD) detachIndependentDisk(disk Disk) error {
+
+	// See if the disk is attached to the VM
+	vmRef, err := disk.AttachedVM()
+	if err != nil {
+		return err
+	}
+	// If the disk is attached to the VM, detach disk from the VM
+	if vmRef != nil {
+
+		vm, err := vcd.client.Client.GetVMByHref(vmRef.HREF)
+		if err != nil {
+			return err
+		}
+
+		// Detach the disk from VM
+		task, err := vm.DetachDisk(&types.DiskAttachOrDetachParams{
+			Disk: &types.Reference{
+				HREF: disk.Disk.HREF,
+			},
+		})
+		if err != nil {
+			return err
+		}
+		err = task.WaitTaskCompletion()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// moved from vapp_test.go
+func verifyNetworkConnectionSection(check *C, actual, desired *types.NetworkConnectionSection) {
+
+	check.Assert(len(actual.NetworkConnection), Equals, len(desired.NetworkConnection))
+	check.Assert(actual.PrimaryNetworkConnectionIndex, Equals, desired.PrimaryNetworkConnectionIndex)
+
+	sort.SliceStable(actual.NetworkConnection, func(i, j int) bool {
+		return actual.NetworkConnection[i].NetworkConnectionIndex <
+			actual.NetworkConnection[j].NetworkConnectionIndex
+	})
+
+	for _, nic := range actual.NetworkConnection {
+		actualNic := actual.NetworkConnection[nic.NetworkConnectionIndex]
+		desiredNic := desired.NetworkConnection[nic.NetworkConnectionIndex]
+
+		check.Assert(actualNic.MACAddress, Not(Equals), "")
+		check.Assert(actualNic.NetworkAdapterType, Not(Equals), "")
+		check.Assert(actualNic.IPAddressAllocationMode, Equals, desiredNic.IPAddressAllocationMode)
+		check.Assert(actualNic.Network, Equals, desiredNic.Network)
+		check.Assert(actualNic.NetworkConnectionIndex, Equals, desiredNic.NetworkConnectionIndex)
+
+		if actualNic.IPAddressAllocationMode != types.IPAllocationModeNone {
+			check.Assert(actualNic.IPAddress, Not(Equals), "")
+		}
+	}
+}
+
+// Ensure vApp is suitable for VM test
+// Some VM tests may fail if vApp is not powered on, so VM tests can call this function to ensure the vApp is suitable for VM tests
+// moved from vm_test.go
+func (vcd *TestVCD) ensureVappIsSuitableForVMTest(vapp VApp) error {
+	status, err := vapp.GetStatus()
+
+	if err != nil {
+		return err
+	}
+
+	// If vApp is not powered on (status = 4), power on vApp
+	if status != types.VAppStatuses[4] {
+		task, err := vapp.PowerOn()
+		if err != nil {
+			return err
+		}
+		err = task.WaitTaskCompletion()
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// Ensure VM is suitable for VM test
+// Please call ensureVappAvailableForVMTest first to power on the vApp because this function cannot handle VM in suspension state due to lack of VM APIs (e.g. discard VM suspend API)
+// Some VM tests may fail if VM is not powered on or powered off, so VM tests can call this function to ensure the VM is suitable for VM tests
+// moved from vm_test.go
+func (vcd *TestVCD) ensureVMIsSuitableForVMTest(vm *VM) error {
+	// if the VM is not powered on (status = 4) or not powered off, wait for the VM power on
+	// wait for around 1 min
+	valid := false
+	for i := 0; i < 6; i++ {
+		status, err := vm.GetStatus()
+		if err != nil {
+			return err
+		}
+
+		// If the VM is powered on (status = 4)
+		if status == types.VAppStatuses[4] {
+			// Prevent affect Test_ChangeMemorySize
+			// because TestVCD.Test_AttachedVMDisk is run before Test_ChangeMemorySize and Test_ChangeMemorySize will fail the test if the VM is powered on,
+			task, err := vm.PowerOff()
+			if err != nil {
+				return err
+			}
+			err = task.WaitTaskCompletion()
+			if err != nil {
+				return err
+			}
+		}
+
+		// If the VM is powered on (status = 4) or powered off (status = 8)
+		if status == types.VAppStatuses[4] || status == types.VAppStatuses[8] {
+			valid = true
+		}
+
+		// If 1st to 5th attempt is completed, sleep 10 seconds and try again
+		// The last attempt will exit this for loop immediately, so no need to sleep
+		if i < 5 {
+			time.Sleep(time.Second * 10)
+		}
+	}
+
+	if !valid {
+		return errors.New("the VM is not powered on or powered off")
+	}
+
+	return nil
+}
+
+// moved from org_test.go
+func doesOrgExist(check *C, vcd *TestVCD) {
+	var org *AdminOrg
+	for i := 0; i < 30; i++ {
+		org, _ = vcd.client.GetAdminOrgByName(TestDeleteOrg)
+		if org == nil {
+			break
+		} else {
+			time.Sleep(time.Second)
+		}
+	}
+	check.Assert(org, IsNil)
+}
+
+// Helper function that creates an external network to be used in other tests
+// moved from externalnetwork_test.go
+func (vcd *TestVCD) testCreateExternalNetwork(testName, networkName, dnsSuffix string) (skippingReason string, externalNetwork *types.ExternalNetwork, task Task, err error) {
+
+	if vcd.skipAdminTests {
+		return fmt.Sprintf(TestRequiresSysAdminPrivileges, testName), externalNetwork, Task{}, nil
+	}
+
+	if vcd.config.VCD.ExternalNetwork == "" {
+		return fmt.Sprintf("%s: External network isn't configured. Test can't proceed", testName), externalNetwork, Task{}, nil
+	}
+
+	if vcd.config.VCD.VimServer == "" {
+		return fmt.Sprintf("%s: Vim server isn't configured. Test can't proceed", testName), externalNetwork, Task{}, nil
+	}
+
+	if vcd.config.VCD.ExternalNetworkPortGroup == "" {
+		return fmt.Sprintf("%s: Port group isn't configured. Test can't proceed", testName), externalNetwork, Task{}, nil
+	}
+
+	if vcd.config.VCD.ExternalNetworkPortGroupType == "" {
+		return fmt.Sprintf("%s: Port group type isn't configured. Test can't proceed", testName), externalNetwork, Task{}, nil
+	}
+
+	virtualCenters, err := QueryVirtualCenters(vcd.client, fmt.Sprintf("name==%s", vcd.config.VCD.VimServer))
+	if err != nil {
+		return "", externalNetwork, Task{}, err
+	}
+	if len(virtualCenters) == 0 {
+		return fmt.Sprintf("No vSphere server found with name '%s'", vcd.config.VCD.VimServer), externalNetwork, Task{}, nil
+	}
+	vimServerHref := virtualCenters[0].HREF
+
+	// Resolve port group info
+	portGroups, err := QueryPortGroups(vcd.client, fmt.Sprintf("name==%s;portgroupType==%s", url.QueryEscape(vcd.config.VCD.ExternalNetworkPortGroup), vcd.config.VCD.ExternalNetworkPortGroupType))
+	if err != nil {
+		return "", externalNetwork, Task{}, err
+	}
+	if len(portGroups) == 0 {
+		return fmt.Sprintf("No port group found with name '%s'", vcd.config.VCD.ExternalNetworkPortGroup), externalNetwork, Task{}, nil
+	}
+	if len(portGroups) > 1 {
+		return fmt.Sprintf("More than one port group found with name '%s'", vcd.config.VCD.ExternalNetworkPortGroup), externalNetwork, Task{}, nil
+	}
+
+	externalNetwork = &types.ExternalNetwork{
+		Name:        networkName,
+		Description: "Test Create External Network",
+		Configuration: &types.NetworkConfiguration{
+			IPScopes: &types.IPScopes{
+				IPScope: []*types.IPScope{&types.IPScope{
+					Gateway:   "192.168.201.1",
+					Netmask:   "255.255.255.0",
+					DNS1:      "192.168.202.253",
+					DNS2:      "192.168.202.254",
+					DNSSuffix: dnsSuffix,
+					IPRanges: &types.IPRanges{
+						IPRange: []*types.IPRange{
+							&types.IPRange{
+								StartAddress: "192.168.201.3",
+								EndAddress:   "192.168.201.100",
+							},
+							&types.IPRange{
+								StartAddress: "192.168.201.105",
+								EndAddress:   "192.168.201.140",
+							},
+						},
+					},
+				}, &types.IPScope{
+					Gateway:   "192.168.231.1",
+					Netmask:   "255.255.255.0",
+					DNS1:      "192.168.232.253",
+					DNS2:      "192.168.232.254",
+					DNSSuffix: dnsSuffix,
+					IPRanges: &types.IPRanges{
+						IPRange: []*types.IPRange{
+							&types.IPRange{
+								StartAddress: "192.168.231.3",
+								EndAddress:   "192.168.231.100",
+							},
+							&types.IPRange{
+								StartAddress: "192.168.231.105",
+								EndAddress:   "192.168.231.140",
+							},
+							&types.IPRange{
+								StartAddress: "192.168.231.145",
+								EndAddress:   "192.168.231.150",
+							},
+						},
+					},
+				},
+				}},
+			FenceMode: "isolated",
+		},
+		VimPortGroupRefs: &types.VimObjectRefs{
+			VimObjectRef: []*types.VimObjectRef{
+				&types.VimObjectRef{
+					VimServerRef: &types.Reference{
+						HREF: vimServerHref,
+					},
+					MoRef:         portGroups[0].MoRef,
+					VimObjectType: vcd.config.VCD.ExternalNetworkPortGroupType,
+				},
+			},
+		},
+	}
+	task, err = CreateExternalNetwork(vcd.client, externalNetwork)
+	return skippingReason, externalNetwork, task, err
+}
+
+// deleteLbServerPoolIfExists is used to cleanup before creation of component. It returns error only if there was
+// other error than govcd.ErrorEntityNotFound
+// moved from lbserverpool_test.go
+func deleteLbServerPoolIfExists(edge EdgeGateway, name string) error {
+	err := edge.DeleteLbServerPoolByName(name)
+	if err != nil && !ContainsNotFound(err) {
+		return err
+	}
+	if err != nil && ContainsNotFound(err) {
+		return nil
+	}
+
+	fmt.Printf("# Removed leftover LB server pool'%s'\n", name)
+	return nil
+}
+
+// deleteLbServiceMonitorIfExists is used to cleanup before creation of component. It returns error only if there was
+// other error than govcd.ErrorEntityNotFound
+// moved from lbservicemonitor_test.go
+func deleteLbServiceMonitorIfExists(edge EdgeGateway, name string) error {
+	err := edge.DeleteLbServiceMonitorByName(name)
+	if err != nil && !ContainsNotFound(err) {
+		return err
+	}
+	if err != nil && ContainsNotFound(err) {
+		return nil
+	}
+
+	fmt.Printf("# Removed leftover LB service monitor'%s'\n", name)
+	return nil
+}
+
+// deleteLbAppProfileIfExists is used to cleanup before creation of component. It returns error only if there was
+// other error than govcd.ErrorEntityNotFound
+// moved from lbappprofile_test.go
+func deleteLbAppProfileIfExists(edge EdgeGateway, name string) error {
+	err := edge.DeleteLbAppProfileByName(name)
+	if err != nil && !ContainsNotFound(err) {
+		return err
+	}
+	if err != nil && ContainsNotFound(err) {
+		return nil
+	}
+
+	fmt.Printf("# Removed leftover LB app profile '%s'\n", name)
+	return nil
+}
+
+// deleteLbAppRuleIfExists is used to cleanup before creation of component. It returns error only if there was
+// other error than govcd.ErrorEntityNotFound
+// moved from lbapprule_test.go
+func deleteLbAppRuleIfExists(edge EdgeGateway, name string) error {
+	err := edge.DeleteLbAppRuleByName(name)
+	if err != nil && !ContainsNotFound(err) {
+		return err
+	}
+	if err != nil && ContainsNotFound(err) {
+		return nil
+	}
+
+	fmt.Printf("# Removed leftover LB app rule '%s'\n", name)
+	return nil
+}
+
+// moved from vm_test.go
+func deleteVapp(vcd *TestVCD, name string) error {
+	vapp, err := vcd.vdc.GetVAppByName(name, true)
+	if err != nil {
+		return fmt.Errorf("error getting vApp: %s", err)
+	}
+	task, _ := vapp.Undeploy()
+	_ = task.WaitTaskCompletion()
+
+	// Detach all Org networks during vApp removal because network removal errors if it happens
+	// very quickly (as the next task) after vApp removal
+	task, _ = vapp.RemoveAllNetworks()
+	err = task.WaitTaskCompletion()
+	if err != nil {
+		return fmt.Errorf("error removing networks from vApp: %s", err)
+	}
+
+	task, err = vapp.Delete()
+	if err != nil {
+		return fmt.Errorf("error deleting vApp: %s", err)
+	}
+	err = task.WaitTaskCompletion()
+	if err != nil {
+		return fmt.Errorf("error waiting for vApp deletion task: %s", err)
+	}
+	return nil
 }
