+++ govcd/entity_test.go
@@ -1,4 +1,4 @@
-// +build api functional catalog org extnetwork vm vdc system user nsxv network ALL
+// +build api functional catalog org extnetwork vm vdc system user nsxv network vapp vm affinity ALL
 
 /*
  * Copyright 2019 VMware, Inc.  All rights reserved.  Licensed under the Apache v2 License.
