+++ govcd/lbserverpool_test.go
@@ -7,10 +7,9 @@
 package govcd
 
 import (
-	"fmt"
+	. "gopkg.in/check.v1"
 
 	"github.com/vmware/go-vcloud-director/v2/types/v56"
-	. "gopkg.in/check.v1"
 )
 
 // Test_LBServerPool tests CRUD methods for load balancer server pool.
@@ -149,19 +148,4 @@ func (vcd *TestVCD) Test_LBServerPool(check *C) {
 
 	_, err = edge.GetLbServerPoolById(createdLbPool.ID)
 	check.Assert(IsNotFound(err), Equals, true)
-}
-
-// deleteLbServerPoolIfExists is used to cleanup before creation of component. It returns error only if there was
-// other error than govcd.ErrorEntityNotFound
-func deleteLbServerPoolIfExists(edge EdgeGateway, name string) error {
-	err := edge.DeleteLbServerPoolByName(name)
-	if err != nil && !ContainsNotFound(err) {
-		return err
-	}
-	if err != nil && ContainsNotFound(err) {
-		return nil
-	}
-
-	fmt.Printf("# Removed leftover LB server pool'%s'\n", name)
-	return nil
 }
